requirements.txt
# NexSupply Platform Dependencies
# Python 3.9+

# Core Framework
streamlit>=1.28.0

# AI/LLM
google-generativeai>=0.3.0

# Data Processing
pandas>=2.0.0

# Visualization
plotly>=5.18.0

# Image Processing
pillow>=10.0.0

# Environment
python-dotenv>=1.0.0
